requirement.txt
amqp==5.3.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.1.0
Flask-Celery==2.4.3
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
numpy==2.2.3
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
